refactor(vscode): share not-found error detection
