+++ packages/kilo-vscode/src/kilo-provider/handlers/question.ts
@@ -7,6 +7,7 @@
  */
 
 import type { KiloClient, QuestionRequest } from "@kilocode/sdk/v2/client"
+import { isNotFoundError } from "./not-found"
 
 export interface QuestionContext {
   readonly client: KiloClient | null
@@ -29,19 +30,6 @@ interface QuestionRecovery {
 }
 
 type QuestionRoute = { kind: "retry"; dir: string } | { kind: "stale" } | { kind: "failed" }
-
-function isNotFoundError(error: unknown): boolean {
-  const record = (value: unknown) =>
-    value && typeof value === "object" ? (value as Record<string, unknown>) : undefined
-  const obj = record(error)
-  if (!obj) return false
-
-  const cause = record(obj.cause)
-  const body = record(cause?.body)
-  return [obj, record(obj.data), cause, body, record(body?.data)].some(
-    (value) => value?.name === "NotFoundError" || value?._tag === "NotFound" || value?.status === 404,
-  )
-}
 
 async function recover(ctx: QuestionContext, requestID: string): Promise<QuestionRoute> {
   const result = await fetchAndSendPendingQuestions(ctx, requestID)
@@ -141,7 +129,7 @@ export async function handleQuestionReply(
     ctx.clearQuestionDirectory(requestID)
     return true
   } catch (error) {
-    const route = isNotFoundError(error) ? await recover(ctx, requestID) : undefined
+    const route = isNotFoundError(error, true) ? await recover(ctx, requestID) : undefined
     if (route?.kind === "stale") return false
     if (route?.kind === "retry" && route.dir !== dir) {
       try {
@@ -179,7 +167,7 @@ export async function handleQuestionReject(
     ctx.clearQuestionDirectory(requestID)
     return true
   } catch (error) {
-    const route = isNotFoundError(error) ? await recover(ctx, requestID) : undefined
+    const route = isNotFoundError(error, true) ? await recover(ctx, requestID) : undefined
     if (route?.kind === "stale") return false
     if (route?.kind === "retry" && route.dir !== dir) {
       try {
